feat: fixes build incompatibilities between packages
TSC builds a format that is not compatible with node resolution
algorithm. Trying to build with tsc a node-compatible format results in
other incompatibilities with how the code is written so I've just
implemented a quick build step with rollup that resolves in a single
file, which is fully compatible with node.